requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2 